requirements.txt
python-telegram-bot==21.4
openai==1.51.0
asyncpg==0.29.0
redis==5.0.8
pydantic==2.8.2
python-dotenv==1.0.1
pandas==2.2.2
XlsxWriter==3.2.0
matplotlib==3.9.0
Pillow==10.4.0
email-validator==2.2.0
bleach==6.1.0
structlog==24.1.0
aiofiles==23.2.1

httpx==0.27.2

aiohttp==3.10.10
